relocate POJO
